+++ NEWS.md
@@ -1,5 +1,7 @@
 # dbplyr (development version)
 
+* Moved argument `auto_index` after `...` in `*_join()` (@mgirlich).
+
 * Removed dependency on assertthat (@mgirlich, #1112).
 
 * Catch unsupported argument `pivot_wider(id_expand = TRUE)` and